Implement setindex for Base.ImmutableDict
The design of Base.ImmutableDict allows the same key to be added again, which
shadows the previously added value. Unfortunately, it still prints out all of
the entries, including repeated keys. I thought this might be confusing, so I
implemented it to remove the value if the key exists and then add the new value.

Closes #216